+++ src/typings/global.d.ts
@@ -1,6 +1,12 @@
 declare global {
   var srfbe: {
     wpVersion: string;
+		postType: string;
+		isShortcutEnabled: boolean;
+		isCaseMatchingEnabled: boolean;
+		isRegexMatchingEnabled: boolean;
+		isCloseModalEnabled: boolean;
+		isSavePostEnabled: boolean;
   };
 }